ergodox_ez/austin: Replace left alt with win/cmd
The left thumb key in question is primarily used as a WM leader. The
reason for keeping it as left alt was convenience on standard keyboards;
however, I unintentionally trained myself to Win when using a friend's
machine. With that hurdle out of the way, Win is undoubtedly the right
choice -- fewer shortcut conflicts, and less wierdness (i.e. in Firefox,
where Alt alone focuses the menu).