ergodox_ez/austin: add gaming layer
The first change required here is to put spacebar on the left thumb, to
support jumping with mouse + wasd. going forward, I suspect there will
be other tweaks to make here.

Would be nice to find a way to express this (at compile time) as a
delta from my base keymap, rather than duping everything... but with the
current way the LAYOUT macros work this doesn't look too feasible.